refactor(session): isolate faction events
