+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Session/Session.Events.cs
@@ -1,6 +1,4 @@
 using System.Collections.Generic;
-using System.Linq;
-using Sandbox.Game;
 using Sandbox.Game.Entities;
 using Sandbox.ModAPI;
 using VRage.Game.ModAPI;
@@ -20,7 +18,8 @@ namespace ShipCoreFramework
 
             if (group.LinkType == GridLinkTypeEnum.Physical)
             {
-                TrackPhysicalGridGroup(group);
+                if (IsServer)
+                    TrackPhysicalGridGroup(group);
                 return;
             }
 
@@ -46,8 +45,11 @@ namespace ShipCoreFramework
 
             if (!IsInitialGroupScan)
             {
-                RefreshPhysicalGroupLinkagesForGrids(tempGridList);
-                gComp.QueueConnectorNetworkRefresh();
+                if (IsServer)
+                {
+                    RefreshPhysicalGroupLinkagesForGrids(tempGridList);
+                    gComp.QueueConnectorNetworkRefresh();
+                }
             }
         }
 
@@ -108,7 +110,8 @@ namespace ShipCoreFramework
 
             if (group.LinkType == GridLinkTypeEnum.Physical)
             {
-                UntrackPhysicalGridGroup(group);
+                if (IsServer)
+                    UntrackPhysicalGridGroup(group);
                 return;
             }
 
@@ -123,95 +126,8 @@ namespace ShipCoreFramework
             group.OnGridRemoved -= gComp.OnGridRemoved;
             gComp.Clean();
 
-            RefreshPhysicalGroupLinkagesForGrids(tempGridList);
-        }
-        
-        private static void FactionStateChanged(MyFactionStateChange action, long fromFactionId, long toFactionId,
-            long factionId, long playerId)
-        {
-            if (!IsServer) return;
-            if (Config.SelectedNoCore == null) return;
-            if (playerId > 0)
-            {
-                var identityFactionId = toFactionId > 0 ? toFactionId : factionId > 0 ? factionId : fromFactionId;
-                PerFactionManager.TrackFactionIdentity(playerId, identityFactionId);
-            }
-
-            if (!IsRelevantFactionStateChange(action)) return;
-            Utils.Log($"FactionStateChanged: {action} from {fromFactionId} to {toFactionId} for faction {factionId} and player {playerId}", 1);
-
-            if (action == MyFactionStateChange.RemoveFaction)
-            {
-                var removedFactionGroups = GetAffectedGroupsForFactionChange(factionId, 0).ToList();
-                PerFactionManager.RemoveFaction(factionId);
-                EnforceOverCapacityForGroups(removedFactionGroups);
-                return;
-            }
-
-            if (action == MyFactionStateChange.FactionMemberAcceptJoin)
-            {
-                var newFactionId = toFactionId > 0 ? toFactionId : factionId;
-                EnforceOverCapacityForGroups(GetAffectedGroupsForFactionChange(newFactionId, playerId));
-                return;
-            }
-
-            var oldFactionId = fromFactionId > 0 ? fromFactionId : factionId;
-            var affectedGroups = GetAffectedGroupsForFactionChange(oldFactionId, playerId).ToList();
-
-            foreach (var comp in affectedGroups
-                         .Where(group => group.OwnerId == playerId)
-                         .ToList())
-            {
-                PerFactionManager.RemoveGridGroup(oldFactionId, comp.ShipCore.SubtypeId);
-            }
-
-            EnforceOverCapacityForGroups(affectedGroups);
-        }
-
-        private static void FactionCreated(long factionId)
-        {
-            if (!IsServer) return;
-            PerFactionManager.TrackFactionMembers(factionId);
-        }
-
-        private static void FactionEdited(long factionId)
-        {
-            if (!IsServer) return;
-            PerFactionManager.TrackFactionMembers(factionId);
-        }
-        
-        private static void SessionReady()
-        {
-            if (Config.SelectedNoCore == null || !IsServer) return;
-            PerFactionManager.InitializeIdentityCache();
-            MyAPIGateway.Session.DamageSystem.RegisterBeforeDamageHandler(-100, CubeGridModifiers.GridCoreDamageHandler);
-            MyExplosions.OnExplosion += CubeGridModifiers.HandleLightningExplosions;
-        }
-
-        private static bool IsRelevantFactionStateChange(MyFactionStateChange action)
-        {
-            return action == MyFactionStateChange.FactionMemberAcceptJoin ||
-                   action == MyFactionStateChange.FactionMemberKick ||
-                   action == MyFactionStateChange.FactionMemberLeave ||
-                   action == MyFactionStateChange.RemoveFaction;
-        }
-
-        private static IEnumerable<GroupComponent> GetAffectedGroupsForFactionChange(long factionId, long playerId)
-        {
-            return GroupDict.Values.Where(group => group.MainCoreComponent != null &&
-                                                   (group.OwnerId == playerId ||
-                                                    factionId > 0 &&
-                                                    group.OwningFaction != null &&
-                                                    group.OwningFaction.FactionId == factionId));
-        }
-
-        private static void EnforceOverCapacityForGroups(IEnumerable<GroupComponent> groups)
-        {
-            foreach (var comp in groups.Where(group => group?.MainCoreComponent != null).Distinct().ToList())
-            {
-                comp.SyncBeaconComponents();
-                comp.RefreshPunishmentState();
-            }
+            if (IsServer)
+                RefreshPhysicalGroupLinkagesForGrids(tempGridList);
         }
     }
 }
